Create code
In the code above, we first define the API key and the API endpoint URL that we will use to fetch the exchange rate data. We then use the fetch() function to send a GET request to the API endpoint and retrieve the response as a JSON object.
We extract the exchange rate for USD to EUR from the response data using the data.rates.EUR property, and log it to the console.
Note that you will need to sign up for an API key from the Open Exchange Rates website and replace YOUR_API_KEY with your actual API key in the code above.